==> 1. nfl betting.ipynb <==
# Scraping NFL game statistics

import pandas as pd
import numpy as np
import requests

import random
import time

#selecting data for seasons 2016-2024
seasons = [str(season) for season in range(2016,2025)]

#three-letter codes for html data
team_abb = ['crd', 'atl', 'rav', 'buf', 'car', 'chi', 'cin', 'cle', 'dal', 'den', 'det', 'gnb', 'htx', 'clt', 'jax', 'kan', 'sdg', 'ram', 'rai', 
           'mia', 'min', 'nwe', 'nor', 'nyg', 'nyj', 'phi', 'pit', 'sea', 'sfo', 'tam', 'oti', 'was']

nfl_df = pd.DataFrame()

#Scrapes html data for each season from 2016-2024
for season in seasons:
    for team in team_abb:
        url = 'https://www.pro-football-reference.com/teams/'+team+'/'+season+'/gamelog'

        print(url)
        
        off_df = pd.read_html(url, header=1, attrs={'id':'gamelog' + season})[0]

        def_df = pd.read_html(url,header=1,attrs = {'id':'gamelog_opp'+season})[0]

        team_df = pd.concat([off_df,def_df], axis=1)

        team_df.insert(loc=0, column = 'Season', value = season)
        team_df.insert(loc=1, column = 'Team', value = team.upper())

        #dataframe stored in 'nfl_df'
        nfl_df = pd.concat([nfl_df,team_df], ignore_index=True) 

        time.sleep(random.randint(4,5))

#saves nfl_df dataframe to .csv file
nfl_df.to_csv('nfl_gamelogs_2016-2024.csv',index=False)

nfl_df.shape
