==> 2. vegas_lines.ipynb <==
# Scraping nfl data in relation to betting lines

import pandas as pd
import numpy as np
import requests

import random
import time

seasons = [str(season) for season in range(2016,2025)]

team_abb = ['crd', 'atl', 'rav', 'buf', 'car', 'chi', 'cin', 'cle', 'dal', 'den', 'det', 'gnb', 'htx', 'clt', 'jax', 'kan', 'sdg', 'ram', 'rai', 
           'mia', 'min', 'nwe', 'nor', 'nyg', 'nyj', 'phi', 'pit', 'sea', 'sfo', 'tam', 'oti', 'was']

veg_df = pd.DataFrame()

for season in seasons:
    for team in team_abb:
        url = 'https://www.pro-football-reference.com/teams/'+team+'/'+season+'_lines.htm'

        print(url)
        
        lines_df = pd.read_html(url, header=0, attrs={'id':'vegas_lines'})[0]

        lines_df.insert(loc=0, column = 'Season', value = season)
        lines_df.insert(loc=1, column = 'Team', value = team.upper())


        veg_df = pd.concat([veg_df,lines_df], ignore_index=True) 

        time.sleep(random.randint(4,5))

veg_df.shape

veg_df.to_csv('vegaslines_2016-2024.csv',index=False)
