==> 3. modifying_gamelogs.ipynb <==
import pandas as pd

nfl_df = pd.read_csv('nfl_gamelogs_2016-2024.csv')

print(nfl_df.shape[1])

#dictionary for new column names
column_names = {'Unnamed: 4':'Result', 'Unnamed: 6': 'Home', 'Tm': 'Pts_F', 'Opp.1': 'Pts_A', 
               'Att.1':'Rsh_Att','Yds.2':'Rsh_Yds', 'Y/A.1':'Rsh_Y/A', 'TD.1':'Rsh_TD', 'Yds.3':'Pnt_Yds'}

#renaming dataframe's select column names
nfl_df = nfl_df.rename(columns = column_names)

#viewing the unique values in the 'Unnamed 3:' column
nfl_df['Unnamed: 3'].unique()

#removing the rows with 'preview' as 'Unnamed 3:' value. This indicates a future game
nfl_df = nfl_df[nfl_df['Unnamed: 3'] =='boxscore']

#copying dataframe to keep progress
nfl_df2= nfl_df

#dropping irrelevant data
nfl_df2 = nfl_df2.drop(nfl_df2.columns[5], axis=1)
nfl_df2 = nfl_df2.drop(nfl_df2.columns[17], axis=1)
nfl_df2 = nfl_df2.drop(nfl_df2.columns[36:46], axis=1)
nfl_df2 = nfl_df2.drop(nfl_df2.columns[42], axis=1)

#copying dataframe to keep progress
nfl_df3=nfl_df2

#marking columns that contain opponent stats
opp_cols = nfl_df3.columns[36:]

#adding prefix 'Opp_' and removing trailing numbers from opponent columns
nfl_df3 = nfl_df3.rename(columns={c: 'Opp_'+c[:-2] for c in nfl_df3.columns if c in opp_cols})

column_names2 = {'Opp_Att.1':'Opp_Rsh_Att','Opp_Yds.2':'Opp_Rsh_Yds', 'Opp_Y/A.1':'Opp_Rsh_Y/A', 'Opp_TD.1':'Opp_Rsh_TD', 'Opp_Yds.3':'Opp_Pnt_Yds', }
nfl_df3 = nfl_df3.rename(columns = column_names2)

nfl_df3.columns

#copying dataframe to keep progress
nfl_df4 = nfl_df3

#mapping string data to binary (1 or 0)
nfl_df4['OT'] = nfl_df4['OT'].apply(lambda x: 1 if x == 'OT' else 0)
nfl_df4['Home'] = nfl_df4['Home'].apply(lambda x: 0 if x == '@' else 1)                                    

#mapping full names to three-letter codes
team_dict = {'Arizona Cardinals':'CRD', 'Atlanta Falcons':'ATL', 'Baltimore Ravens':'RAV', 'Buffalo Bills':'BUF', 'Carolina Panthers':'CAR',
             'Chicago Bears':'CHI', 'Cincinnati Bengals':'CIN', 'Cleveland Browns':'CLE', 'Dallas Cowboys':'DAL', 'Denver Broncos':'DEN',
             'Detroit Lions':'DET', 'Green Bay Packers':'GNB', 'Houston Texans':'HTX', 'Indianapolis Colts':'CLT', 'Jacksonville Jaguars':'JAX',
             'Kansas City Chiefs':'KAN', 'Los Angeles Chargers':'SDG', 'Los Angeles Rams':'RAM', 'Las Vegas Raiders':'RAI', 'Miami Dolphins':'MIA',
             'Minnesota Vikings':'MIN', 'New England Patriots':'NWE', 'New Orleans Saints':'NOR', 'New York Giants':'NYG', 'New York Jets':'NYJ',
             'Philadelphia Eagles':'PHI', 'Pittsburgh Steelers':'PIT', 'St. Louis Rams':'RAM', 'San Diego Chargers':'SDG',
             'San Francisco 49ers':'SFO', 'Seattle Seahawks':'SEA', 'Tampa Bay Buccaneers':'TAM', 'Tennessee Titans':'OTI',
             'Washington Commanders':'WAS', 'Washington Football Team':'WAS', 'Washington Redskins':'WAS'}
nfl_df4 = nfl_df4.replace({'Opp': team_dict})

nfl_df4.to_csv('nfl_gamelogs_2016-2024_modified.csv',index=False)

nfl_df4.shape
