==> 4. merging.ipynb <==
import pandas as pd

nfl_df = pd.read_csv('nfl_gamelogs_2016-2024_modified.csv')

veg_df = pd.read_csv('vegaslines_2016-2024.csv')

veg_df = veg_df.rename(columns = {'G#':'G', 'Over/Under':'Total'})

veg_df = veg_df.drop(veg_df.columns[6:],axis=1)

#veg_df.iloc[:,0].unique()

#removing playoff games from data
#seasons before 2021 had 16 games whereas seasons after have 17 games
veg_df = veg_df.query('(Season < 2021 and G < 17) or (Season >= 2021 and G < 18)')

veg_df

veg_df['Home'] = veg_df['Opp'].apply(lambda x: 0 if x[0] == '@' else 1)

veg_df['Opp'] = veg_df['Opp'].apply(lambda x: x[1:] if x[0] == '@' else x)

veg_df

veg_df2 = veg_df[['Season', 'Team', 'G', 'Home', 'Opp', 'Spread', 'Total']]

veg_df2.columns

veg_df2

abbr_dict = {'OAK':'RAI', 'LVR':'RAI', 'STL':'RAM', 'LAR':'RAM', 'LAC':'SDG', 'IND':'CLT',
             'HOU':'HTX', 'BAL':'RAV', 'ARI':'CRD', 'TEN':'OTI'}
veg_df2 = veg_df2.replace({'Opp': abbr_dict})

veg_df2

nfl_df

veg_df2.shape

nfl_df.shape

#merge nfl game statistics with betting lines based on these 4 keys 
merged_df = pd.merge(nfl_df,veg_df2, on=['Season', 'Team', 'Home', 'Opp'])

merged_df.to_csv('nfl_betting.csv', index=False)
