==> 5. trends.ipynb <==
import pandas as pd

seasons = [str(season) for season in range(2016,2025)]

nfl_df = pd.read_csv('nfl_betting.csv')

nfl_df

#finding trends in data
#finding how many games with certain betting totals end up hitting 'under', and average home team pass yards per game
condition = 'Total.isin([46.5, 47.0, 47.5]) and Home == 1'
result = 'Pts_F + Pts_A < Total'

print(f"Condition = '{condition}' and Result = '{result}' ({min(seasons)} - {max(seasons)}):\n")

cond_total = 0
res_total = 0
stat_total = 0

for season in seasons:
    season_df = nfl_df.query(f'Season == {season}')
    s_count = len(season_df)
    
    cond_df = season_df.query(condition)
    c_count = len(cond_df)

    result_df = cond_df.query(result)
    r_count = len(result_df)

    stat_count = result_df['Yds'].sum()
    
    if c_count > 0:
        print(f' {season} {r_count / c_count:.2%} ({r_count} of {c_count}) Avg Pass Yds: {stat_count/r_count:.5}')
    else:
        print('nothing')

    cond_total += c_count
    res_total += r_count
    stat_total += stat_count

print(f"\nTotal: {res_total / cond_total:.2%} ({res_total} of {cond_total}) Avg Pass Yds: {stat_total/res_total:.5}")

#finding trends based on previous data
#find underdogs that cover the spread after losing by 20+ pts
prev_condition = 'Pts_A - Pts_F >= 20'
curr_condition = 'Pts_F + Pts_A < Total'
cols = ['Team', 'G']
result = 'Pts_F + Spread > Pts_A'

print(f"Previous Condition = '{prev_condition}', Current Condition = '{curr_condition}', and Result = '{result}' ({min(seasons)} - {max(seasons)}):\n")

cond_total = 0
res_total = 0

for season in seasons:
    season_df = nfl_df.query(f'Season == {season}')
    s_count = len(season_df)

    prev_cond_df = season_df.query(prev_condition)
    #create tuples of instances where previous condition was met
    prev_tuples = prev_cond_df[cols].to_records(index=False).tolist()

    #find the next game after 20+ pt loss
    curr_tuples = [(team, game+1) for team,game in prev_tuples]

    #find current df based on current tuples
    curr_df = season_df[season_df[cols].apply(func = tuple,axis=1).isin(curr_tuples)]

    cond_df = curr_df.query(curr_condition)
    c_count = len(cond_df)

    result_df = cond_df.query(result)
    r_count = len(result_df)

    if c_count > 0:
        print(f' {season} {r_count / c_count:.2%} ({r_count} of {c_count})')
    else:
        print('nothing')

    cond_total += c_count
    res_total += r_count

print(f"\nTotal: {res_total / cond_total:.2%} ({res_total} of {cond_total})")
